pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Antonio Pereira <anthonyesany0712@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.10"
black = "^23.12.1"
django = "^4.2.5"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"
pytest-django = "^4.5.2"
django-debug-toolbar = "^4.2.0"
autopep8 = "^1.6.0"
isort = "^5.10.1"
flake8 = "^6.1.0"
python-decouple = "^3.8"
lxml = "^5.1.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.1"
factory-boy = "^3.3.0"

[tool.poetry.group.dev.dependencies]
mypy = {git = "https://github.com/python/mypy.git"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
